se agregan campos, se agrega reporte ConsultasVacaciones, se agrega un campo de limite acumulacion para vacacionRelacionLaboral

+++ WebAppTH/bd.webappth.entidades/Negocio/SolicitudVacaciones.cs
@@ -48,7 +48,7 @@ namespace bd.webappth.entidades.Negocio
         public int? IdEmpleadoReemplazo { get; set; }
 
         [NotMapped]
-        public int DiasVacaciones { get; set; }
+        public decimal DiasVacaciones { get; set; }
 
         [Display(Name = "Requiere reemplazo")]
         public bool RequiereReemplazo { get; set; }
